+++ examples/zia_admin_user/admin_user_department_scope.tf
@@ -1,6 +1,6 @@
 resource "zia_admin_users" "john_smith" {
   login_name                      = "john.smith@acme.com"
-  user_name                       = "John Smith"
+  username                       = "John Smith"
   email                           = "john.smith@acme.com"
   is_password_login_allowed       = true
   password                        = "AeQ9E5w8B$"
@@ -11,12 +11,10 @@ resource "zia_admin_users" "john_smith" {
   role {
     id = data.zia_admin_roles.super_admin.id
   }
-  admin_scope {
-    type = "DEPARTMENT"
-    scope_entities {
-      id = [data.zia_department_management.engineering.id]
+  admin_scope_type = "DEPARTMENT"
+    admin_scope_entities {
+        id = [ data.zia_department_management.engineering.id, data.zia_department_management.sales.id ]
     }
-  }
 }
 
 data "zia_admin_roles" "super_admin" {
@@ -26,3 +24,8 @@ data "zia_admin_roles" "super_admin" {
 data "zia_department_management" "engineering" {
   name = "Engineering"
 }
+
+data "zia_department_management" "sales" {
+  name = "Sales"
+}
+